Always run "apt-get update"
When only leaf images are rebuilt, the Packages index might be outdated
and refer to packages that no longer exist in the archive. Always
running "apt-get update" before installing packages avoids this.

Also remove contents of /var/lib/apt/lists as we no longer use it and it
allows to keep the images slightly smaller.