Bump base-10 image to clang 5
The base-10 image is currently using clang 4, which has major compatibility issues with the stdlib
of GCC 7 in C++17 mode, mostly revolving around std::string_view. This makes it impossible to use
the clang in this image for compiling Dune.

This patch resolves the issue by simply bumping clang to version 5.